Bug 1266343 - Use an empty mozconfig from the source directory. r=chmanchester
When configure unit tests use an empty mozconfig, instead of creating an
empty temporary file, use an empty mozconfig from the source directory.

+++ python/mozbuild/mozbuild/test/configure/common.py
@@ -170,9 +170,13 @@ class BaseConfigureTest(unittest.TestCase):
         if not out:
             out = StringIO()
 
-        fh, mozconfig_path = tempfile.mkstemp()
-        os.write(fh, mozconfig)
-        os.close(fh)
+        if mozconfig:
+            fh, mozconfig_path = tempfile.mkstemp()
+            os.write(fh, mozconfig)
+            os.close(fh)
+        else:
+            mozconfig_path = os.path.join(os.path.dirname(__file__), 'data',
+                                          'empty_mozconfig')
 
         try:
             environ = dict(
@@ -192,4 +196,5 @@ class BaseConfigureTest(unittest.TestCase):
 
             return sandbox
         finally:
-            os.remove(mozconfig_path)
+            if mozconfig:
+                os.remove(mozconfig_path)
